Change method of work environment detection
It does not rely on hostname and so anymore. It instead asks whether it
is a work environment or not, and retain this information in later runs.

There is also now a method for setting up the work environment in
non-interactive environments, such as Codespaces and Gitpod, which is by
using the environment variable DOTFILES_IS_WORK=true.