Make homebrew manage chezmoi installation again
It turns out that otherwise, it may be more difficult to handle the
bash and zsh completions of it.

But this time, we also inverted the order of upgrade in the full-upgrade
script, so that homebrew packages gets upgraded before the dotfiles
itself.

This can potentially fix a problem of chezmoi refusing to apply the
dotfiles due to being outdated.