Move home folder to subdirectory (#32)
This moves all files which are actually mapped to the home folder with chezmoi to a subfolder of the repository called home.

This gives me peace of mind for adding more files to the repo without worrying about polluting my actual home folder. This way, things are much more organized.

This also includes a migration script, so users who still are using the old TOML format for configuration, or have their chezmoi not pointing to the home subfolder will be prompted for fixing it by reinitializing chezmoi. The migration also handles the deletion of old configuration. Deleting the whole configuration is needed because of https://github.com/twpayne/chezmoi/issues/1431.

References https://github.com/twpayne/chezmoi/issues/891