Use code shim to rule them all
Now, it's the responsible for falling back to other editors if code is
not available. This was causing chezmoi not to select code as the editor
in fresh install.

Also, it handles the presence of both code in Windows and Linux in case
someone is using WSL. This is useful because I use code installed in
Linux to aid testing of VS Code extensions. There is now the
USE_CODE_FROM_LINUX environment variable that can be used to control
this behavior.

Finally, it handles the case where I hit 'chezmoi edit' and expect the
command not to wait for the closure of VS Code, as explained at
https://github.com/twpayne/chezmoi/issues/1068.