Install direnv with chezmoiexternal
As now https://github.com/direnv/direnv/pull/845 was merged and we
finally have first-class support for .env files. The move to
chezmoiexternal was due to 2 things: because I want the new version fast
and because I make it indenpendent of the OS, allowing it to be used on
devcontainers as well.