Fix direnv not loading env variables automatically
This issue is happening because direnv is now being installed to
~/.local/bin instead of the default system-wide location. And, this dir
only gets added to PATH after sourcing .profile, which happens after
loading direnv.

Now, as direnv will always be installed, even in devcontainers, we call
it using the full path instead of just direnv. But anyway, to prevent
possible issues in unkown environments, it also checks if the
~/.local/bin/direnv exist and is executable.

Also changes emulate zsh -c to eval, because I don't know what are the
benefits of using emulate anyway. And brew also uses eval.

/cc @mrbernardi