Move the installation of apt packages to earlier
There is a rare situation where jq (which is a requirement for running
templates in the dotfiles) is only installed at the user level (e.g.
when installed by homebrew). This means that home's setup-prerequisites.sh
thinks that jq is good while it is actually not.

When that happens, the chezmoi execution fails when applying the root
files. Moving the installation of jq to earlier will ensure that it is
installed by apt before the root files are applied.

/cc @mrbernardi