Ensure pre-requisites are met for rootmoi as well
There is a rare situation where jq (which is a requirement for running
templates in the dotfiles) is only installed at the user level (e.g.
when installed by homebrew). This means that home's setup-prerequisites.sh
thinks that jq is good while it is actually not. When that happens, the
chezmoi execution fails when applying the root files.

The first approach tried in 7417ae4618504f948d5c0740bbd3018727c2b386 did
not work because chezmoi actually creates the apt repository files which
are required before installing the apt packages.

This new approach simply symlinks the install-prerequisites script from
home to root.

/cc @mrbernardi